requirements.txt
python-chess
requests
cairosvg
